requirements.txt
firebase-admin==5.3.0
google-api-core==2.8.2
google-api-python-client==2.61.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
numpy==1.22.4
requests==2.28.0


